+++ v2/processor_test.go
@@ -2,7 +2,6 @@ package v2_test
 
 import (
 	"context"
-	"fmt"
 	"testing"
 	"time"
 
@@ -53,74 +52,6 @@ func ExampleProcessor() {
 	cancel()
 }
 
-var _ v2.MessageSettler = &fakeSettler{}
-
-type fakeSettler struct {
-	AbandonCalled    int
-	CompleteCalled   int
-	DeadLetterCalled int
-	DeferCalled      int
-	RenewCalled      int
-}
-
-func (f *fakeSettler) AbandonMessage(ctx context.Context, message *azservicebus.ReceivedMessage, options *azservicebus.AbandonMessageOptions) error {
-	f.AbandonCalled++
-	return nil
-}
-
-func (f *fakeSettler) CompleteMessage(ctx context.Context, message *azservicebus.ReceivedMessage, options *azservicebus.CompleteMessageOptions) error {
-	f.CompleteCalled++
-	return nil
-}
-
-func (f *fakeSettler) DeadLetterMessage(ctx context.Context, message *azservicebus.ReceivedMessage, options *azservicebus.DeadLetterOptions) error {
-	f.DeadLetterCalled++
-	return nil
-}
-
-func (f *fakeSettler) DeferMessage(ctx context.Context, message *azservicebus.ReceivedMessage, options *azservicebus.DeferMessageOptions) error {
-	f.DeferCalled++
-	return nil
-}
-
-func (f *fakeSettler) RenewMessageLock(ctx context.Context, message *azservicebus.ReceivedMessage, options *azservicebus.RenewMessageLockOptions) error {
-	f.RenewCalled++
-	return nil
-}
-
-type fakeReceiver struct {
-	// outcomes to verify
-	ReceiveCalls []int // array of maxMessage value passed to receive calls in the lifetime of the fake receiver
-
-	// configure fake
-	SetupReceiveError     error
-	SetupReceivedMessages chan *azservicebus.ReceivedMessage
-	*fakeSettler
-	SetupMaxReceiveCalls int
-}
-
-func (f *fakeReceiver) ReceiveMessages(ctx context.Context, maxMessages int, options *azservicebus.ReceiveMessagesOptions) ([]*azservicebus.ReceivedMessage, error) {
-	f.ReceiveCalls = append(f.ReceiveCalls, maxMessages)
-	if maxMessages == 0 && len(f.SetupReceivedMessages) > 0 {
-		return nil, nil
-	}
-	var result []*azservicebus.ReceivedMessage
-	for msg := range f.SetupReceivedMessages {
-		result = append(result, msg)
-		if len(result) == maxMessages || len(f.SetupReceivedMessages) == 0 {
-			fmt.Println("no more messages, break out")
-			break
-		}
-	}
-
-	// return an error if we request more messages than there are available.
-	if len(f.ReceiveCalls) >= f.SetupMaxReceiveCalls {
-		return result, fmt.Errorf("max receive calls exceeded")
-	}
-
-	return result, f.SetupReceiveError
-}
-
 func TestNewProcessor_DefaultsToMaxConcurrency1(t *testing.T) {
 	a := require.New(t)
 	messages := make(chan *azservicebus.ReceivedMessage, 1)
@@ -156,20 +87,6 @@ func TestNewProcessor_CanSetMaxConcurrency(t *testing.T) {
 	a.EqualError(err, "max receive calls exceeded")
 	a.Equal(1, len(rcv.ReceiveCalls), "there should be 1 entry in the ReceiveCalls array")
 	a.Equal(10, rcv.ReceiveCalls[0], "the processor should have used max concurrency of 10")
-}
-
-func messagesChannel(messageCount int) chan *azservicebus.ReceivedMessage {
-	messages := make(chan *azservicebus.ReceivedMessage, messageCount)
-	for i := 0; i < messageCount; i++ {
-		messages <- &azservicebus.ReceivedMessage{}
-	}
-	return messages
-}
-
-func enqueueCount(q chan *azservicebus.ReceivedMessage, messageCount int) {
-	for i := 0; i < messageCount; i++ {
-		q <- &azservicebus.ReceivedMessage{}
-	}
 }
 
 func TestNewProcessor_Interval(t *testing.T) {
